Global update

+++ src/main/java/org/ngelmakproject/web/rest/dto/ActiveChannel.java
@@ -7,5 +7,5 @@ public record ActiveChannel(
 		String avatar,
 		String banner,
 		String description,
-		Integer postCount) {
+		Long postCount) {
 }
